Finish Create Account Module, DataUpload Blockchaon Module

+++ src/main/java/com/example/Blockchain/UserInfo/UserInfoService.java
@@ -1,15 +1,20 @@
 package com.example.Blockchain.UserInfo;
 
 import com.example.Blockchain.CommentModel.Encoder_MD5;
+import com.example.Blockchain.web3Info.nodeService;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.web3j.crypto.CipherException;
 import org.web3j.protocol.Web3j;
-import org.web3j.protocol.core.DefaultBlockParameter;
-import org.web3j.protocol.core.methods.response.EthGetBalance;
 
 import javax.transaction.Transactional;
+import java.io.IOException;
+import java.security.InvalidAlgorithmParameterException;
+import java.security.NoSuchAlgorithmException;
+import java.security.NoSuchProviderException;
+import java.sql.Timestamp;
 import java.util.*;
 
 @Service
@@ -18,13 +23,15 @@ public class UserInfoService {
     private final UserInfoRepository userInfoRepository;
     private final Encoder_MD5 encoder_md5;
     private final Web3j web3j;
+    private final nodeService nodeService;
 
 
     @Autowired
-    public UserInfoService(UserInfoRepository userInfoRepository, Encoder_MD5 encoder_md5, Web3j web3j) {
+    public UserInfoService(UserInfoRepository userInfoRepository, Encoder_MD5 encoder_md5, Web3j web3j, com.example.Blockchain.web3Info.nodeService nodeService) {
         this.userInfoRepository = userInfoRepository;
         this.encoder_md5 = encoder_md5;
         this.web3j = web3j;
+        this.nodeService = nodeService;
     }
 
     public List<UserInfo> getUserInfo(){
@@ -32,18 +39,22 @@ public class UserInfoService {
     }
 
 
-    public void addUser(UserInfo userInfo){
+    public void addUser(UserInfo userInfo) throws CipherException, InvalidAlgorithmParameterException, NoSuchAlgorithmException, NoSuchProviderException, IOException {
         boolean exists = userInfoRepository.existsById(userInfo.getUserAccount());
         if (exists){
             throw new IllegalStateException("userAccount:" + userInfo.getUserAccount() + "已被使用");
         }else {
+            //紀錄創建時間以及初始化最終修改時間
+            Long datetime = System.currentTimeMillis();
+            Timestamp timestamp = new Timestamp(datetime);
+            userInfo.setCreatedTime(timestamp);
+            userInfo.setUpdatedTime(timestamp);
             //MD5加密
             userInfo.setUserPassword(encoder_md5.encodeMD5(userInfo.getUserPassword()));
+            //取得區塊鏈錢包
+            String userAddress = "0x" + nodeService.createAccount(userInfo.getUserPassword());
+            userInfo.setUserAddress(userAddress);
             userInfoRepository.save(userInfo);
-
-//            EthGetBalance result = web3j.ethGetBalance("0x6e61a11b9a8c9a3fd1904919ef0668cbed170fd7",DefaultBlockParameter.valueOf(s)).send();
-
-
         }
 
     }
@@ -53,14 +64,12 @@ public class UserInfoService {
                 () -> new IllegalStateException("userAccount:" + userAccount + "不存在")
         );
 
-//        if ( !Objects.equals(userInfo.getUserPassword(),encoder_md5.encodeMD5(userPassword))){
-//            throw new IllegalStateException("密碼錯誤");
-//
-//        }else{
-//            userInfoRepository.deleteById(userAccount);
-//        }
-        userInfoRepository.deleteById(userAccount);
+        if ( !Objects.equals(userInfo.getUserPassword(),encoder_md5.encodeMD5(userPassword))){
+            throw new IllegalStateException("密碼錯誤");
 
+        }else{
+            userInfoRepository.deleteById(userAccount);
+        }
     }
 
     @Transactional
@@ -70,15 +79,18 @@ public class UserInfoService {
         );
 
         //新密碼不為空，不與先前密碼相同
-        if ( !Objects.equals(userInfo.getUserPassword(),originalUserPassword)){
+        if ( !Objects.equals(userInfo.getUserPassword(),encoder_md5.encodeMD5(originalUserPassword))){
             throw new IllegalStateException("密碼錯誤");
         }else if (newUserPassword == null || newUserPassword.length() <= 0){
             throw new IllegalStateException("新密碼不得為空");
         }else if (Objects.equals(userInfo.getUserPassword(),newUserPassword)){
             throw new IllegalStateException("新密碼不得與舊密碼相同");
         }else {
-            userInfo.setUserPassword(newUserPassword);
-
+            userInfo.setUserPassword(encoder_md5.encodeMD5(newUserPassword));
+            //紀錄修改時間
+            Long datetime = System.currentTimeMillis();
+            Timestamp timestamp = new Timestamp(datetime);
+            userInfo.setUpdatedTime(timestamp);
         }
 //        if (Objects.equals(userInfo.getUserPassword(),originalUserPassword)
 //                && newUserPassword != null
@@ -94,10 +106,45 @@ public class UserInfoService {
         );
 
         if (userAddress == null || userAddress.length() <= 0){
-            throw new IllegalStateException("暱稱不得為空");
+            throw new IllegalStateException("Address不得為空");
         }else{
             userInfo.setUserAddress(userAddress);
 
+        }
+    }
+
+    //個人資料update，沒修改就用原本的(serviceName,agenciesName,status,.....)
+    @Transactional
+    public void updateInfo(String userAccount, String serviceName,String agenciesName,String status){
+        UserInfo userInfo = userInfoRepository.findById(userAccount).orElseThrow(
+                () -> new IllegalStateException("userAccount:" + userAccount + "不存在")
+        );
+
+        if (agenciesName == null || agenciesName.length() <= 0){
+            throw new IllegalStateException("名稱不得為空");
+        }else{
+            userInfo.setServiceName(serviceName);
+            userInfo.setAgenciesName(agenciesName);
+            userInfo.setStatus(status);
+            //紀錄修改時間
+            Long datetime = System.currentTimeMillis();
+            Timestamp timestamp = new Timestamp(datetime);
+            userInfo.setUpdatedTime(timestamp);
+
+        }
+    }
+
+    @Transactional
+    public String exchangeAddress(String userAccount,String userPassword){
+        UserInfo userInfo = userInfoRepository.findById(userAccount).orElseThrow(
+                () -> new IllegalStateException("userAccount:" + userAccount + "不存在")
+        );
+
+        if ( !Objects.equals(userInfo.getUserPassword(),encoder_md5.encodeMD5(userPassword))){
+            throw new IllegalStateException("密碼錯誤");
+
+        }else{
+            return userInfo.getUserAddress();
         }
     }
